+++ packages/xchain-avax/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.3.5 (2023-11-03)
+
+## Update
+
+- EIP1559 params
+
+# v0.3.4 (2023-11-02)
+
+## Update
+
+- Estimations can be done with data provider
+
 # v0.3.3 (2023-10-26)
 
 ## Update