+++ packages/xchain-midgard-query/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.1.5 (2023-11-05)
+
+## Update
+
+- getDecimalForAsset do not override midgard response
+
+# v0.1.4 (2023-11-04)
+
+## Update
+
+- getDecimalForAsset can handle synth assets
+
 # v0.1.3 (2023-10-06)
 
 ## Update